Fixed Bugs in File IO, improved Code Style

+++ src/game/constants.rs
@@ -13,7 +13,7 @@ pub const PLAYER_SPEED: f32 = 4.0;
 /// The pickup distance for coins
 pub const COLLECTIBLE_PICKUP_DISTANCE: f32 = 0.1;
 /// The UI Font Size for main UI elements
-pub const _UI_FONT_SIZE: f32 = 30 as f32;
+pub const _UI_FONT_SIZE: f32 = 30.;
 /// The speed a dead player ascends to heaven at
 pub const DEAD_PLAYER_ASCEND_SPEED: f32 = 3.0;
 /// The speed a dead player zooms while ascending
